requirements.txt
numpy~=1.19.5
pytesseract~=0.3.8
matplotlib~=3.4.3
opencv-python~=4.5.4.58
Pillow~=8.4.0
pdf2image~=1.16.0
pandas~=1.3.4
setuptools~=57.0.0
PyPDF2~=1.26.0
torch~=1.10.0
PyYAML~=6.0
seaborn~=0.11.2
scipy~=1.7.1
torchvision~=0.11.1
tqdm~=4.62.3
thop~=0.0.31.post2005241907
requests~=2.26.0
roboflow~=0.1.8
ipython~=7.28.0
wxPython~=4.1.1

------------------------------------------------------------------------------------------------------------------------

# Clone repository
git clone https://github.com/ultralytics/yolov5

# Change directory
%cd yolov5
git reset --hard 886f1c03d839575afecb059accf74296fad395b6

# Install dependencies
pip install -qr requirements.txt

